Ensure habits that are due are shown on habits page
Updated the backend to fetch only habits due on the selected day. This data is retrieved from either a join between habits and habit_progress or habits and habit_instances. The retrieval only goes as far as habits due within the next 7 days, starting from the habit creation date. This limit can be adjusted by passing in a different number as the daysAhead parameter for the generateIntervalInstances function